+++ src/engine-internal/config/selectors.ts
@@ -1,36 +1,46 @@
 /**
- * @file Config state slice selectors, used to get Config parameters directly from global state
+ * @file {@link https://github.com/CianciarusoCataldo/modular-engine modular-engine} `config` slice selectors
+ *
+ * @see https://cianciarusocataldo.github.io/modular-engine/docs
  *
  * @author Cataldo Cianciaruso <https://github.com/CianciarusoCataldo>
  *
  * @copyright Cataldo Cianciaruso 2022
  */
 
-import { ModularEngineConfigState, ModularEngineGlobalState } from "modular-engine-types";
+import {
+  ModularEngineConfigState,
+  ModularEngineGlobalState,
+} from "modular-engine-types";
 
-import { createModularSelector } from "../../utils";
+import { createModularEngineSelector } from "../../helpers/utils";
 
 import defaultConfig from "./initial-state";
 
 /**
- * Returns all shared config parameters, saved into the Redux state
+ * Returns all  {@link https://github.com/CianciarusoCataldo/modular-engine modular-engine} shared config parameters, saved into the `config` state
+ *
+ * @see https://cianciarusocataldo.github.io/modular-engine/docs
  *
  * @author Cataldo Cianciaruso <https://github.com/CianciarusoCataldo>
  *
  * @copyright Cataldo Cianciaruso 2022
  */
-export const getConfig = (state?: ModularEngineGlobalState): ModularEngineConfigState =>
-  state.config || defaultConfig;
+export const getConfig = (
+  state?: ModularEngineGlobalState
+): ModularEngineConfigState => state.config || defaultConfig;
 
 /**
- * Returns `appName` parameter
+ * Returns {@link https://github.com/CianciarusoCataldo/modular-engine modular-engine} config `appName` parameter
+ *
+ * @see https://cianciarusocataldo.github.io/modular-engine/docs
  *
  * @author Cataldo Cianciaruso <https://github.com/CianciarusoCataldo>
  *
  * @copyright Cataldo Cianciaruso 2022
  *
  */
-export const getAppName = createModularSelector(
+export const getAppName = createModularEngineSelector(
   getConfig,
   ({ appName }) => appName
 );