+++ src/index.ts
@@ -1,8 +1,9 @@
 export {
-  createModularAction,
-  createModularReducer,
-  createModularSelector,
-} from "./utils";
+  createModularEngineAction,
+  createModularEngineReducer,
+  createModularEngineSelector,
+  createModularEnginePlugin,
+} from "./helpers/utils";
 
 export { getAppName, getConfig } from "./engine-internal/config/selectors";
 export { engineInitCompleted } from "./engine-internal/core/actions";