+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Versions
 
+- [0.4.0](#040)
 - [0.3.0](#030)
 - [0.2.0](#020)
 - [0.1.0](#010)
@@ -14,10 +15,18 @@
 
 <br>
 
+### 0.4.0
+
+- New naming convention, every helper function name now contains `engine` word to better differentiate them
+- Refactoring
+- Tests updated
+
+<br>
+
 ### 0.3.0
 
 - `createModularEnginePlugin` helper function, to easily create a modular-engine plugin with the new format
-- types updated
+- Types updated
 
 <br>
 
@@ -31,4 +40,4 @@
 
 ### 0.1.0
 
-- Modular util functions extracted from main lib
+- Modular-engine util functions extracted from main lib